+++ vg/vg_snap.h
@@ -25,7 +25,7 @@ struct ag_menu;
 struct ag_menu_item;
 
 __BEGIN_DECLS
-void		 VG_SnapPoint(struct vg *, double *, double *);
+void		 VG_SnapPoint(struct vg *, float *, float *);
 __inline__ void	 VG_SetSnapMode(struct vg *, enum vg_snap_mode);
 __inline__ void	 VG_DrawGrid(struct vg *);
 struct ag_toolbar *VG_SnapToolbar(void *, struct vg *, enum ag_toolbar_type);
